// lib/types.ts
export type ShippingAddress = {
    fullName: string;
    addressLine1: string;
    addressLine2?: string;
    city: string;
    postalCode: string;
    country: string;
    phoneNumber?: string;
};

export type Order = {
  id: string;
  userId: string;
  items: {
    productId: string;
    name: string;
    quantity: number;
    priceAtPurchase: number;
    color: string;
    size?: string;
    imageUrl?: string;
  }[];
  totalAmount: number;
  status: 'pending' | 'paid' | 'shipped' | 'delivered' | 'cancelled' | 'refunded';
  orderDate: Date; // Changed to Date to be more robust
  shippingAddress?: ShippingAddress;
};
// ...
export type ProductSales = {
  productId: string;
  name: string;
  quantitySold: number;
};

export type SalesAnalytics = {
  totalRevenue: number;
  revenueAmine: number;
  revenueMaryem: number;
  totalOrdersDelivered: number;
  totalItemsSold: number;
  totalCustomers: number; // New
  monthlySales: {
    [key: string]: { // key is 'YYYY-MM'
      revenue: number;
      orders: number;
    }
  };
  topSellingProducts: ProductSales[];
  lastUpdated: any; // Firestore Timestamp
};